% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHospitals.R
\docType{data}
\name{AllHospitals}
\alias{AllHospitals}
\title{Excerpt of data from all German Hospitals over the years 2015 to 2019.}
\format{
A \code{tibble} with 11326 rows and 40 columns:
\describe{
  \item{\code{year}}{When was the data reported?}
  \item{\code{ID}}{Overall ID code of this hospital.}
  \item{\code{idGlobalHospitalIdentifier}}{Overall ID code of the hospital but
      according to the primary key in the database.}
  \item{\code{idHospitalOperator}}{ID of the owner / operator of the hospital.}
  \item{\code{HospitalOperatorName}}{The name of the owner /operator.}
  \item{\code{type}}{What kind of status has the owner of the company (e.g.,
      public or private).}
  \item{\code{idHospital}}{ID of the individual data entry (one hospital has
      several of these IDs over the years.)}
  \item{\code{HospitalName}}{The name of the hospital.}
  \item{\code{ikNumber}}{The official ID number of the hospital. Be aware that
      this can be error-prone and changing over the years.}
  \item{\code{locationNumberHospital}}{Number of the location of the hospital.}
  \item{\code{locationNumberSite}}{Number of the location of the site of the
      hospital.}
  \item{\code{locationNumberOverall}}{Location number for this hospital, regardeless
      if it's the overall number or the site number.}
  \item{\code{phoneHospital}}{Contact phone of the hospital.}
  \item{\code{urlHospital}}{URL to the website of the hospital.}
  \item{\code{urlAdditionalInformation}}{Website with additional information
      regarding the hospital.}
  \item{\code{academicTeachingHospital}}{Is this hospital a clinic where new
      health professionals are getting their education?}
  \item{\code{universityClinic}}{Is it a university clinic?}
  \item{\code{universityName}}{If it's a teaching hospital but no university,
      what is the name of the university connected to it?}
  \item{\code{psychiatricHospital}}{Does the hospital have a psychiatric unit?}
  \item{\code{psychiatricDutyToSupply}}{Is this hospital obligated to treat
      psychiatric cases?}
  \item{\code{quantityBeds}}{How many beds are available in this hospital?}
  \item{\code{quantityCasesFull}}{How many inpatient cases were treated in this
     hospital?}
  \item{\code{quantityCasesPartial}}{How many day care inpatient cases were
      treated in this hospital?}
  \item{\code{quantityCasesOutpatient}}{How many outpatient cases were treated in this
     hospital?}
  \item{\code{ExternalAttendingDoctors}}{How many external doctors are working
      in this hospital?}
  \item{\code{Doctors}}{How many doctors (without counting externals) are working
      in this hospital?}
  \item{\code{AttendingDoctors}}{How many of these doctors (without counting externals)
      are attending doctors?}
  \item{\code{Nurses}}{How many nurses are working in this hospital?}
  \item{\code{tariffWeeklyWorkingHoursDoctors}}{What are the mean working hours for
      doctors in this hospital?}
  \item{\code{tariffWeeklyWorkingHoursNurses}}{What are the mean working hours for
      nurses in this hospital?}
  \item{\code{street}}{Part of the address of this hospital.}
  \item{\code{housenumber}}{Part of the address of this hospital.}
  \item{\code{zip}}{Part of the address of this hospital.}
  \item{\code{city}}{Part of the address of this hospital.}
  \item{\code{district}}{Part of the address of this hospital.}
  \item{\code{state}}{Part of the address of this hospital.}
  \item{\code{country}}{Part of the address of this hospital.}
  \item{\code{URL}}{Another link to the URL of this hospital.}
  \item{\code{lat}}{Latitude regarding this hospital.}
  \item{\code{lon}}{Longitude regarding this hospital.}
}
}
\usage{
AllHospitals
}
\description{
This dataset contains the names, addresses, coordinates and several key
variables from all German Hospitals that are involved in the treatment of
statutory health insurance patients.
}
\keyword{datasets}
