% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unemploymentCountsAge.R
\docType{data}
\name{unemploymentCountsAge}
\alias{unemploymentCountsAge}
\title{Unemployment numbers regarding every county in Germany from 2015 to 2020 with a focus on age.}
\format{
A \code{tibble} with 9378 rows and five columns:
\describe{
  \item{\code{year}}{A string with values "2015", "2016", "2017", "2018", "2019"
      and "2020", indicating to which year the data belongs. Every year is
      reported on the last day of the year, i.e., the 31st of December each
      year.}
  \item{\code{AGS}}{This character string comprises the "Amtliche Gemeindeschluessel"
      (AGS) for each county, where the first and second character represents
      the state, the third character the district and the fourth and fifth
      character represents the county.}
  \item{\code{name}}{A character value with the name of the respective county.}
  \item{\code{ageGroup}}{A character value indicating the age group the data belongs to.
      There are only three categories, and two of these are overlapping:

      \itemize{
          \item{15_to_less_than_20_years}
          \item{15_to_less_than_25_years}
          \item{55_to_less_than_65_years}
      }
  }
  \item{\code{count}}{An integer value with the count of unemployed people.}
}
}
\source{
Obtained from the \emph{Regionalstatistik} data base:

    \href{https://www.regionalstatistik.de/genesis/online?operation=find&suchanweisung_language=de&query=13211-02-05-4}{Table "13211-02-05-4 - Arbeitslose nach ausgewählten Personengruppen sowie Arbeitslosenquoten - Jahresdurchschnitt"}

    \emph{Source:} Arbeitsmarktstatistik der Bundesagentur für Arbeit (BA).

    \emph{Copyright:} "Statistische Aemter des Bundes und der Laender, Deutschland, 2021.
    Dieses Werk ist lizenziert unter der Datenlizenz Deutschland - Namensnennung -
    Version 2.0."

    \emph{Licence:} \url{www.govdata.de/dl-de/by-2-0}

    Obtained on the 17.08.2021 at 19:38:04.
}
\usage{
unemploymentCountsAge
}
\description{
This dataset contains unemployment numbers from all German counties over the
years 2015 to 2020, with a sole focus on younger and older persons.
}
\keyword{datasets}
