% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbgbaExtraData.R
\docType{package}
\name{qbgbaExtraData}
\alias{qbgbaExtraData}
\title{qbgbaExtraData: A data package complementing the "Qualitaetsberichte" for
German hospitals.}
\description{
This package aims at bundling not only key variables from the reports published by
the "Gemeinsamer Bundesausschuss" (GBA) but also several supplementary datasets that
can be used in conjunction with these.
}
